renderer: limit labels on x-axis  (#249)

+++ modules/ref_fb_module/include/ref_fb_module/classifier_fb_impl.h
@@ -57,7 +57,7 @@ private:
     BlockReaderPtr linearReader;
 
     size_t blockSize;
-    Float classCount;
+    size_t classCount;
     Float inputDeltaTicks;
     Float inputResolution;
